backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-real-estate-agent-backend"
version = "0.1.0"
description = "FastAPI backend for AI Real Estate Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Private"}
authors = [
    {name = "AI Real Estate Agent Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "pypdf>=3.17.0",
    "PyPDF2>=3.0.1",
    "reportlab>=4.0.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "google-generativeai>=0.3.0",
    "stripe>=7.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "cuid2>=2.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
